refactor: 유저 로그아웃 및 회원탈퇴 구현
